Modify metaconfig instead of overwriting it

+++ source/Public/Get-GuestConfigurationPackageComplianceStatus.ps1
@@ -62,7 +62,18 @@ function Get-GuestConfigurationPackageComplianceStatus
 
             # Set LCM settings to force load powershell module.
             $metaConfigPath = Join-Path -Path $PackagePath -ChildPath "$packageName.metaconfig.json"
-            "{""debugMode"":""ForceModuleImport""}" | Out-File $metaConfigPath -Encoding ascii
+            # If metaconfig already exists, append
+            if (Test-Path $metaConfigPath)
+            {
+                $metaConfigObject = Get-Content -Path $metaConfigPath | ConvertFrom-Json -AsHashTable
+                $metaConfigObject["debugMode"] = "ForceModuleImport"
+                $metaConfigObject | ConvertTo-Json | Out-File $metaConfigPath -Encoding ascii -Force
+            }
+            else
+            {
+                "{""debugMode"":""ForceModuleImport""}" | Out-File $metaConfigPath -Encoding ascii
+            }
+
             Set-DscLocalConfigurationManager -ConfigurationName $packageName -Path $PackagePath -Verbose:$verbose
